Update DDL.sql and main.py for charset and query
The commit involves updating the DDL.sql file to alter the charset of the 'words' and 'word_count' tables to utf8mb4 and to general_ci collation. This caters for internationalization and emojis. Additionally, changes have been made in the main.py file under the src/scraper directory. The select statement was updated to order by a random function rather than by the id attribute. This provides a more random sample for the scraper.